clean up PyListTemplate and PyItemTemplate from pyscript module
